Optimize CUDA Transpose3DKernel by handling more elements in one thread.

Take Transpose([3072,64,512], [0,21]) as example (which is from one of our models):

in A100, the kernel takes:
- FP16: before the change, 631.52us, after the change, 293.79us, x2.15 faster
- FP32: before the change, 650.62us, after the change, 566.94us, x1.15 faster

in V100:
- FP16: before the change, 727.17us, after the change, 518.30us, x1.4 faster
- FP32: before the change, 1.08ms, after the change, 967.17us, x1.12 faster

Also run more perf tests on different input tensor size, no perf regression observed. The optimization works better on FP16 than FP32, and large tensors.